tests/requirements-test.txt
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
requests-mock>=1.9.0
fakeredis>=2.19.0
